Added remove item fron transaction
